+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.9.9 [2016-09-01]
+
+### Features
+
+- Added support for setting call media mode: p2p and mix
+- Added call control API for audio recording
+- Added camera selection in the demo app
+
+
 ## 0.9.7 [2016-04-28]
 
 ### Features